Update formsTest to use UserJob properly
This updates the test to be saving import mappings in the unit test, rather than in the Mapping
